+++ field_friend/__init__.py
@@ -1,4 +1,4 @@
-from . import automations, hardware, interface, log_configuration, navigation, vision
+from . import automations, hardware, interface, localization, log_configuration, vision
 from .system import System
 
-__all__ = ['automations', 'hardware', 'interface', 'navigation', 'vision', 'log_configuration', 'System']
+__all__ = ['automations', 'hardware', 'interface', 'localization', 'vision', 'log_configuration', 'System']
