+++ field_friend/interface/components/leaflet_map.py
@@ -8,8 +8,7 @@ import rosys.geometry
 from nicegui import app, events, ui
 from nicegui.elements.leaflet_layers import GenericLayer, TileLayer
 
-from ...automations import Field
-from ...navigation.geo_point import GeoPoint
+from ...localization.geo_point import GeoPoint
 from .key_controls import KeyControls
 
 if TYPE_CHECKING:
